fix more circular imports
