+++ src/format/markdown/format-markdown.ts
@@ -1,28 +1,13 @@
 /*
-* format-markdown.ts
-*
-* Copyright (C) 2020-2022 Posit Software, PBC
-*
-*/
+ * format-markdown.ts
+ *
+ * Copyright (C) 2020-2022 Posit Software, PBC
+ */
 
 import { kOutputDivs, kVariant } from "../../config/constants.ts";
 import { Format } from "../../config/types.ts";
 import { createFormat, plaintextFormat } from "../formats-shared.ts";
-
-export const kGfmCommonmarkExtensions = [
-  "+autolink_bare_uris",
-  "+emoji",
-  "+footnotes",
-  "+gfm_auto_identifiers",
-  "+pipe_tables",
-  "+strikeout",
-  "+task_lists",
-  "+tex_math_dollars",
-];
-
-export const kGfmCommonmarkVariant = kGfmCommonmarkExtensions.join("");
-
-export const kGfmCommonmarkFormat = `commonmark${kGfmCommonmarkVariant}`;
+import { kGfmCommonmarkVariant } from "./format-markdown-consts.ts";
 
 export function requiresShortcodeUnescapePostprocessor(markdown: string) {
   return markdown.includes("{{{<");
